# openapi-gpt-action.yaml
openapi: 3.1.0
info:
  title: NOMADS HRRR Wrapper
  version: 0.1.0
servers:
  - url: https://YOUR-RENDER-URL.onrender.com
paths:
  /healthz:
    get:
      operationId: healthz
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                additionalProperties: true

  /debug/hrrr_inventory:
    get:
      operationId: debugHrrrInventory
      summary: Inspect wgrib2 inventory lines for one HRRR forecast hour
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, minimum: -90, maximum: 90 }
        - name: lon
          in: query
          required: true
          schema: { type: number, minimum: -180, maximum: 180 }
        - name: fhr
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 48, default: 0 }
        - name: run_utc
          in: query
          required: false
          schema: { type: string }
        - name: force_refresh
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Inventory debug output
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      source: { type: string }
                      model: { type: string }
                      run_utc: { type: string }
                      fhr: { type: integer }
                      nomads_url: { type: string }
                    additionalProperties: true
                  inventory:
                    type: array
                    items: { type: string }
                additionalProperties: true

  /hrrr/point:
    get:
      operationId: hrrrPoint
      summary: HRRR point time series extracted from NOMADS GRIB2
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, minimum: -90, maximum: 90 }
        - name: lon
          in: query
          required: true
          schema: { type: number, minimum: -180, maximum: 180 }
        - name: fhrs
          in: query
          required: false
          schema: { type: string, default: "0,1,2,3,4,5" }
        - name: run_utc
          in: query
          required: false
          schema: { type: string }
        - name: units
          in: query
          required: false
          schema: { type: string, enum: [si, knots], default: knots }
        - name: force_refresh
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: HRRR point time series JSON
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      source: { type: string }
                      model: { type: string }
                      run_utc: { type: string }
                      cycle_hour: { type: integer }
                      units: { type: string }
                      notes:
                        type: array
                        items: { type: string }
                    additionalProperties: true
                  location:
                    type: object
                    properties:
                      lat: { type: number }
                      lon: { type: number }
                    additionalProperties: true
                  hourly:
                    type: object
                    properties:
                      time:
                        type: array
                        items: { type: string }
                      u10:
                        type: array
                        items: { type: number, nullable: true }
                      v10:
                        type: array
                        items: { type: number, nullable: true }
                      wind_speed_10m:
                        type: array
                        items: { type: number, nullable: true }
                      wind_dir_10m:
                        type: array
                        items: { type: number, nullable: true }
                      gust_10m:
                        type: array
                        items: { type: number, nullable: true }
                      mslp:
                        type: array
                        items: { type: number, nullable: true }
                      u925:
                        type: array
                        items: { type: number, nullable: true }
                      v925:
                        type: array
                        items: { type: number, nullable: true }
                      u850:
                        type: array
                        items: { type: number, nullable: true }
                      v850:
                        type: array
                        items: { type: number, nullable: true }
                      z850:
                        type: array
                        items: { type: number, nullable: true }
                      cape:
                        type: array
                        items: { type: number, nullable: true }
                      cin:
                        type: array
                        items: { type: number, nullable: true }
                      lifted_index:
                        type: array
                        items: { type: number, nullable: true }
                      pbl_height:
                        type: array
                        items: { type: number, nullable: true }
                    additionalProperties: true
                  debug:
                    type: object
                    properties:
                      nomads_urls:
                        type: array
                        items: { type: string }
                    additionalProperties: true
                additionalProperties: true